middleware auth

+++ routes/web.php
@@ -26,28 +26,28 @@ Route::get('/', [ListingController::class, 'index']);
 
 
 // GET NEW LISTING FORM
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 
 // POST NEW LISTING
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // SHOW EDIT FORM
-Route::get('listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // GET SINGLE LISTING
 Route::get("/listings/{listing}", [ListingController::class, 'show']);
 
 // UPDATE SINGLE LISTING
-Route::put("/listings/{listing}", [ListingController::class, 'update']);
+Route::put("/listings/{listing}", [ListingController::class, 'update'])->middleware('auth');
 
 // DELETE SINGLE LISTING
-Route::delete("/listings/{listing}", [ListingController::class, 'destroy']);
+Route::delete("/listings/{listing}", [ListingController::class, 'destroy'])->middleware('auth');
 
 
 
 // GET REGISTER PAGE
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // CREATE NEW USER
 Route::post('/users/create', [UserController::class, 'store']);
@@ -56,7 +56,7 @@ Route::post('/users/create', [UserController::class, 'store']);
 Route::post('/logout', [UserController::class, 'logout']);
 
 //SHOW LOGIN PAGE
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 //LOGIN USER
 Route::post('/users/login', [UserController::class, 'authenticate']);
